Rename autofill.ts to ai.ts
